Move leaderboard endpoint to submission view
